virtio/console: split port_io into platform-specific modules
Replace the single port_io.rs file with a port_io/ directory containing mod.rs (shared traits and types), unix.rs, and windows.rs. The public API is unchanged; all platform-specific implementations are now isolated in their respective modules instead of being interleaved with #[cfg] annotations throughout a single file.
