--- src/devices/src/virtio/console/port_io.rs
@@ -1,564 +1,0 @@
-#[cfg(unix)]
-use libc::{
-    F_GETFL, F_SETFL, O_NONBLOCK, STDERR_FILENO, STDIN_FILENO, STDOUT_FILENO, TIOCGWINSZ, fcntl,
-};
-use log::Level;
-#[cfg(unix)]
-use nix::errno::Errno;
-#[cfg(unix)]
-use nix::ioctl_read_bad;
-#[cfg(unix)]
-use nix::poll::{PollFd, PollFlags, PollTimeout, poll};
-#[cfg(unix)]
-use nix::unistd::{dup, isatty};
-use std::fs::File;
-#[cfg(windows)]
-use std::io::IsTerminal;
-use std::io::{self, ErrorKind};
-#[cfg(windows)]
-use std::mem::MaybeUninit;
-#[cfg(unix)]
-use std::os::fd::{AsFd, AsRawFd, BorrowedFd, OwnedFd, RawFd};
-#[cfg(windows)]
-use std::os::windows::io::{AsRawHandle, BorrowedHandle, OwnedHandle};
-use utils::eventfd::EFD_NONBLOCK;
-use utils::eventfd::EventFd;
-#[cfg(windows)]
-use utils::windows::AsRawFd;
-use vm_memory::bitmap::Bitmap;
-use vm_memory::{VolatileMemoryError, VolatileSlice, WriteVolatile};
-#[cfg(windows)]
-use windows_sys::Win32::{
-    Foundation::FALSE,
-    Storage::FileSystem::{ReadFile, WriteFile},
-    System::{
-        Console::{CONSOLE_SCREEN_BUFFER_INFO, GetConsoleScreenBufferInfo},
-        Threading::{INFINITE, WaitForMultipleObjects, WaitForSingleObject},
-    },
-};
-
-pub trait PortInput {
-    fn read_volatile(&mut self, buf: &mut VolatileSlice) -> Result<usize, io::Error>;
-
-    fn wait_until_readable(&self, stopfd: Option<&EventFd>);
-}
-
-pub trait PortOutput {
-    fn write_volatile(&mut self, buf: &VolatileSlice) -> Result<usize, io::Error>;
-
-    fn wait_until_writable(&self);
-}
-
-/// Terminal properties associated with this port
-pub trait PortTerminalProperties: Send + Sync {
-    fn get_win_size(&self) -> (u16, u16);
-}
-
-#[cfg(unix)]
-pub fn stdin() -> Result<Box<dyn PortInput + Send>, nix::Error> {
-    let fd = dup_raw_fd_into_owned(STDIN_FILENO)?;
-    make_non_blocking(&fd)?;
-    Ok(Box::new(PortInputFd(fd)))
-}
-
-#[cfg(unix)]
-pub fn input_to_raw_fd_dup(fd: RawFd) -> Result<Box<dyn PortInput + Send>, nix::Error> {
-    let fd = dup_raw_fd_into_owned(fd)?;
-    make_non_blocking(&fd)?;
-    Ok(Box::new(PortInputFd(fd)))
-}
-
-#[cfg(windows)]
-pub fn input_to_handle_dup(
-    handle: *mut core::ffi::c_void,
-) -> io::Result<Box<dyn PortInput + Send>> {
-    Ok(Box::new(PortInputHandle(dup_handle(handle)?)))
-}
-
-#[cfg(unix)]
-pub fn stdout() -> Result<Box<dyn PortOutput + Send>, nix::Error> {
-    output_to_raw_fd_dup(STDOUT_FILENO)
-}
-
-#[cfg(unix)]
-pub fn stderr() -> Result<Box<dyn PortOutput + Send>, nix::Error> {
-    output_to_raw_fd_dup(STDERR_FILENO)
-}
-
-#[cfg(unix)]
-pub fn term_fd(
-    term_fd: RawFd,
-) -> Result<Box<dyn PortTerminalProperties + Send + Sync>, nix::Error> {
-    let fd = dup_raw_fd_into_owned(term_fd)?;
-    assert!(
-        isatty(&fd).is_ok_and(|v| v),
-        "Expected fd {fd:?}, to be a tty, to query the window size!"
-    );
-    Ok(Box::new(PortTerminalPropertiesFd(fd)))
-}
-
-#[cfg(windows)]
-pub fn term_handle(
-    handle: *mut core::ffi::c_void,
-) -> io::Result<Box<dyn PortTerminalProperties + Send + Sync>> {
-    assert!(
-        unsafe { BorrowedHandle::borrow_raw(handle).is_terminal() },
-        "Expected handle {handle:?}, to be a tty, to query the window size!"
-    );
-    let handle = dup_handle(handle)?;
-    Ok(Box::new(PortTerminalPropertiesHandle(handle)))
-}
-
-pub fn term_fixed_size(width: u16, height: u16) -> Box<dyn PortTerminalProperties + Send + Sync> {
-    Box::new(PortTerminalPropertiesFixed((width, height)))
-}
-
-#[cfg(unix)]
-pub fn input_empty() -> Result<Box<dyn PortInput + Send>, nix::Error> {
-    Ok(Box::new(PortInputEmpty {}))
-}
-
-#[cfg(target_os = "windows")]
-pub fn input_empty() -> Result<Box<dyn PortInput + Send>, io::Error> {
-    Ok(Box::new(PortInputEmpty {}))
-}
-
-#[cfg(unix)]
-pub fn output_file(file: File) -> Result<Box<dyn PortOutput + Send>, nix::Error> {
-    output_to_raw_fd_dup(file.as_raw_fd())
-}
-
-#[cfg(target_os = "windows")]
-pub fn output_file(file: File) -> Result<Box<dyn PortOutput + Send>, io::Error> {
-    output_to_handle_dup(file.as_raw_handle())
-}
-
-#[cfg(unix)]
-pub fn output_to_raw_fd_dup(fd: RawFd) -> Result<Box<dyn PortOutput + Send>, nix::Error> {
-    let fd = dup_raw_fd_into_owned(fd)?;
-    make_non_blocking(&fd)?;
-    Ok(Box::new(PortOutputFd(fd)))
-}
-
-#[cfg(windows)]
-pub fn output_to_handle_dup(
-    handle: *mut core::ffi::c_void,
-) -> Result<Box<dyn PortOutput + Send>, io::Error> {
-    // We skip make_non_blocking() on Windows and rely on the background
-    // tx_thread to safely execute a blocking write.
-    let owned_handle = dup_handle(handle)?;
-    Ok(Box::new(PortOutputHandle(owned_handle)))
-}
-
-pub fn output_to_log_as_err() -> Box<dyn PortOutput + Send> {
-    Box::new(PortOutputLog::new())
-}
-
-#[cfg(unix)]
-struct PortInputFd(OwnedFd);
-
-#[cfg(unix)]
-impl AsRawFd for PortInputFd {
-    fn as_raw_fd(&self) -> RawFd {
-        self.0.as_raw_fd()
-    }
-}
-
-#[cfg(unix)]
-impl PortInput for PortInputFd {
-    fn read_volatile(&mut self, buf: &mut VolatileSlice) -> io::Result<usize> {
-        // This source code is copied from vm-memory, except it fixes an issue, where
-        // the original code would does not handle handle EWOULDBLOCK
-
-        let fd = self.as_raw_fd();
-        let guard = buf.ptr_guard_mut();
-
-        let dst = guard.as_ptr().cast::<libc::c_void>();
-
-        // SAFETY: We got a valid file descriptor from `AsRawFd`. The memory pointed to by `dst` is
-        // valid for writes of length `buf.len() by the invariants upheld by the constructor
-        // of `VolatileSlice`.
-        let bytes_read = unsafe { libc::read(fd, dst, buf.len()) };
-
-        if bytes_read < 0 {
-            let err = std::io::Error::last_os_error();
-            if err.kind() != ErrorKind::WouldBlock {
-                // We don't know if a partial read might have happened, so mark everything as dirty
-                buf.bitmap().mark_dirty(0, buf.len());
-            }
-
-            Err(err)
-        } else {
-            let bytes_read = bytes_read.try_into().unwrap();
-            buf.bitmap().mark_dirty(0, bytes_read);
-            Ok(bytes_read)
-        }
-    }
-
-    fn wait_until_readable(&self, stopfd: Option<&EventFd>) {
-        let mut poll_fds = Vec::new();
-        poll_fds.push(PollFd::new(self.0.as_fd(), PollFlags::POLLIN));
-        if let Some(stopfd) = stopfd {
-            // SAFETY: we trust stopfd won't go away to avoid a dup call here.
-            let borrowed_fd = unsafe { BorrowedFd::borrow_raw(stopfd.as_raw_fd()) };
-            poll_fds.push(PollFd::new(borrowed_fd, PollFlags::POLLIN));
-        }
-        poll(&mut poll_fds, PollTimeout::NONE).expect("Failed to poll");
-    }
-}
-
-#[cfg(target_os = "windows")]
-struct PortInputHandle(OwnedHandle);
-
-#[cfg(target_os = "windows")]
-impl AsRawHandle for PortInputHandle {
-    fn as_raw_handle(&self) -> RawHandle {
-        self.0.as_raw_handle()
-    }
-}
-
-#[cfg(target_os = "windows")]
-impl PortInput for PortInputHandle {
-    fn read_volatile(&mut self, buf: &mut VolatileSlice) -> io::Result<usize> {
-        let len = u32::try_from(buf.len()).map_err(|_| {
-            io::Error::new(ErrorKind::InvalidInput, "buffer length exceeds u32::MAX")
-        })?;
-        let mut bytes_read: u32 = 0;
-        let ret = unsafe {
-            ReadFile(
-                self.as_raw_handle(),
-                buf.ptr_guard_mut().as_ptr(),
-                len,
-                &mut bytes_read,
-                std::ptr::null_mut(),
-            )
-        };
-        if ret == 0 {
-            let err = io::Error::last_os_error();
-            if err.kind() == ErrorKind::BrokenPipe {
-                return Ok(0);
-            }
-            if err.kind() != ErrorKind::WouldBlock {
-                // We don't know if a partial read might have happened, so mark everything as dirty
-                buf.bitmap().mark_dirty(0, buf.len());
-            }
-            Err(err)
-        } else {
-            let n = bytes_read as usize;
-            buf.bitmap().mark_dirty(0, n);
-            Ok(n)
-        }
-    }
-
-    fn wait_until_readable(&self, stopfd: Option<&EventFd>) {
-        let mut handles = vec![self.as_raw_handle()];
-        if let Some(s) = stopfd {
-            handles.push(s.as_raw_fd());
-        }
-        wait_for_handles(&handles);
-    }
-}
-
-#[cfg(unix)]
-struct PortOutputFd(OwnedFd);
-
-#[cfg(unix)]
-impl AsRawFd for PortOutputFd {
-    fn as_raw_fd(&self) -> RawFd {
-        self.0.as_raw_fd()
-    }
-}
-
-#[cfg(unix)]
-impl PortOutput for PortOutputFd {
-    fn write_volatile(&mut self, buf: &VolatileSlice) -> Result<usize, io::Error> {
-        self.0.write_volatile(buf).map_err(|e| match e {
-            VolatileMemoryError::IOError(e) => e,
-            e => {
-                log::error!("Unsuported error from write_volatile: {e:?}");
-                io::Error::other(e)
-            }
-        })
-    }
-
-    fn wait_until_writable(&self) {
-        let mut poll_fds = [PollFd::new(self.0.as_fd(), PollFlags::POLLOUT)];
-        poll(&mut poll_fds, PollTimeout::NONE).expect("Failed to poll");
-    }
-}
-
-#[cfg(target_os = "windows")]
-struct PortOutputHandle(OwnedHandle);
-
-#[cfg(target_os = "windows")]
-impl AsRawHandle for PortOutputHandle {
-    fn as_raw_handle(&self) -> RawHandle {
-        self.0.as_raw_handle()
-    }
-}
-
-#[cfg(target_os = "windows")]
-impl PortOutput for PortOutputHandle {
-    fn write_volatile(&mut self, buf: &VolatileSlice) -> io::Result<usize> {
-        let len = u32::try_from(buf.len()).map_err(|_| {
-            io::Error::new(ErrorKind::InvalidInput, "buffer length exceeds u32::MAX")
-        })?;
-        let mut bytes_written: u32 = 0;
-        let ret = unsafe {
-            WriteFile(
-                self.as_raw_handle(),
-                buf.ptr_guard().as_ptr(),
-                len,
-                &mut bytes_written,
-                std::ptr::null_mut(),
-            )
-        };
-        if ret == 0 {
-            Err(io::Error::last_os_error())
-        } else {
-            Ok(bytes_written as usize)
-        }
-    }
-
-    fn wait_until_writable(&self) {
-        // Because WriteFile is blocking, `write_volatile` will natively pause the
-        // thread until space is available. It will never return io::ErrorKind::WouldBlock.
-        // Therefore, `process_tx` will never invoke this function on Windows.
-    }
-}
-
-#[cfg(unix)]
-fn dup_raw_fd_into_owned(raw_fd: RawFd) -> Result<OwnedFd, nix::Error> {
-    // SAFETY: if raw_fd is invalid the `dup` call below will fail
-    let borrowed_fd = unsafe { BorrowedFd::borrow_raw(raw_fd) };
-    let fd = dup(borrowed_fd)?;
-    Ok(fd)
-}
-
-#[cfg(unix)]
-fn make_non_blocking(as_rw_fd: &impl AsRawFd) -> Result<(), nix::Error> {
-    let fd = as_rw_fd.as_raw_fd();
-    unsafe {
-        let flags = fcntl(fd, F_GETFL, 0);
-        if flags < 0 {
-            return Err(Errno::last());
-        }
-
-        if fcntl(fd, F_SETFL, flags | O_NONBLOCK) < 0 {
-            return Err(Errno::last());
-        }
-    }
-    Ok(())
-}
-
-#[cfg(windows)]
-fn dup_handle(raw: *mut core::ffi::c_void) -> Result<OwnedHandle, io::Error> {
-    let borrowed = unsafe { BorrowedHandle::borrow_raw(raw) };
-    borrowed.try_clone_to_owned()
-}
-
-/// Block until at least one of the given Windows HANDLEs becomes signaled.
-#[cfg(windows)]
-fn wait_for_handles(handles: &[*mut core::ffi::c_void]) {
-    match handles.len() {
-        0 => std::thread::sleep(std::time::Duration::MAX),
-        1 => unsafe {
-            WaitForSingleObject(handles[0], INFINITE);
-        },
-        n => unsafe {
-            WaitForMultipleObjects(n as u32, handles.as_ptr(), FALSE, INFINITE);
-        },
-    }
-}
-
-// Utility to relay log from the VM (the kernel boot log and messages from init)
-// to the rust log
-#[derive(Default)]
-pub struct PortOutputLog {
-    buf: Vec<u8>,
-}
-
-impl PortOutputLog {
-    const FORCE_FLUSH_TRESHOLD: usize = 512;
-    const LOG_TARGET: &'static str = "init_or_kernel";
-
-    fn new() -> Self {
-        Self::default()
-    }
-
-    fn force_flush(&mut self) {
-        log::log!(target: PortOutputLog::LOG_TARGET, Level::Error, "[missing newline]{}", String::from_utf8_lossy(&self.buf));
-        self.buf.clear();
-    }
-}
-
-impl PortOutput for PortOutputLog {
-    fn write_volatile(&mut self, buf: &VolatileSlice) -> Result<usize, io::Error> {
-        self.buf.write_volatile(buf).map_err(io::Error::other)?;
-
-        let mut start = 0;
-        for (i, ch) in self.buf.iter().cloned().enumerate() {
-            if ch == b'\n' {
-                log::log!(target: PortOutputLog::LOG_TARGET, Level::Error, "{}", String::from_utf8_lossy(&self.buf[start..i]));
-                start = i + 1;
-            }
-        }
-        self.buf.drain(0..start);
-        // Make sure to not grow the internal buffer forever!
-        if self.buf.len() > PortOutputLog::FORCE_FLUSH_TRESHOLD {
-            self.force_flush()
-        }
-        Ok(buf.len())
-    }
-
-    fn wait_until_writable(&self) {}
-}
-
-#[cfg(unix)]
-pub struct PortInputSigInt {
-    sigint_evt: EventFd,
-}
-
-#[cfg(unix)]
-impl PortInputSigInt {
-    pub fn new() -> Self {
-        PortInputSigInt {
-            sigint_evt: EventFd::new(EFD_NONBLOCK)
-                .expect("Failed to create EventFd for SIGINT signaling"),
-        }
-    }
-
-    pub fn sigint_evt(&self) -> &EventFd {
-        &self.sigint_evt
-    }
-}
-
-#[cfg(unix)]
-impl Default for PortInputSigInt {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[cfg(unix)]
-impl PortInput for PortInputSigInt {
-    fn read_volatile(&mut self, buf: &mut VolatileSlice) -> Result<usize, io::Error> {
-        self.sigint_evt.read()?;
-        log::trace!("SIGINT received");
-        buf.copy_from(&[3u8]); //ASCII 'ETX' -> generates SIGINIT in a terminal
-        Ok(1)
-    }
-
-    #[cfg(unix)]
-    fn wait_until_readable(&self, stopfd: Option<&EventFd>) {
-        let mut poll_fds = Vec::with_capacity(2);
-        // SAFETY: we trust sigint_evt won't go away to avoid a dup call here.
-        let sigint_bfd = unsafe { BorrowedFd::borrow_raw(self.sigint_evt.as_raw_fd()) };
-        poll_fds.push(PollFd::new(sigint_bfd, PollFlags::POLLIN));
-        if let Some(stopfd) = stopfd {
-            // SAFETY: we trust stopfd won't go away to avoid a dup call here.
-            let stop_bfd = unsafe { BorrowedFd::borrow_raw(stopfd.as_raw_fd()) };
-            poll_fds.push(PollFd::new(stop_bfd, PollFlags::POLLIN));
-        }
-
-        poll(&mut poll_fds, PollTimeout::NONE).expect("Failed to poll");
-    }
-}
-
-pub struct PortInputEmpty {}
-
-impl PortInputEmpty {
-    pub fn new() -> Self {
-        PortInputEmpty {}
-    }
-}
-
-impl Default for PortInputEmpty {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl PortInput for PortInputEmpty {
-    fn read_volatile(&mut self, _buf: &mut VolatileSlice) -> Result<usize, io::Error> {
-        Ok(0)
-    }
-
-    #[cfg(unix)]
-    fn wait_until_readable(&self, stopfd: Option<&EventFd>) {
-        if let Some(stopfd) = stopfd {
-            // SAFETY: we trust stopfd won't go away to avoid a dup call here.
-            let borrowed_fd = unsafe { BorrowedFd::borrow_raw(stopfd.as_raw_fd()) };
-            let mut poll_fds = [PollFd::new(borrowed_fd, PollFlags::POLLIN)];
-            poll(&mut poll_fds, PollTimeout::NONE).expect("Failed to poll");
-        } else {
-            std::thread::sleep(std::time::Duration::MAX);
-        }
-    }
-
-    #[cfg(windows)]
-    fn wait_until_readable(&self, stopfd: Option<&EventFd>) {
-        let handles: Vec<_> = stopfd.iter().map(|s| s.as_raw_fd()).collect();
-        wait_for_handles(&handles);
-    }
-}
-
-struct PortTerminalPropertiesFixed((u16, u16));
-
-impl PortTerminalProperties for PortTerminalPropertiesFixed {
-    fn get_win_size(&self) -> (u16, u16) {
-        self.0
-    }
-}
-
-#[cfg(unix)]
-struct PortTerminalPropertiesFd(OwnedFd);
-
-#[cfg(unix)]
-impl PortTerminalProperties for PortTerminalPropertiesFd {
-    fn get_win_size(&self) -> (u16, u16) {
-        let mut ws: WS = WS::default();
-
-        if let Err(err) = unsafe { tiocgwinsz(self.0.as_raw_fd(), &mut ws) } {
-            error!("Couldn't get terminal dimensions: {err}");
-            return (0, 0);
-        }
-        (ws.cols, ws.rows)
-    }
-}
-
-#[cfg(target_os = "windows")]
-struct PortTerminalPropertiesHandle(OwnedHandle);
-
-#[cfg(target_os = "windows")]
-impl PortTerminalProperties for PortTerminalPropertiesHandle {
-    fn get_win_size(&self) -> (u16, u16) {
-        let mut info = MaybeUninit::<CONSOLE_SCREEN_BUFFER_INFO>::uninit();
-        let ret = unsafe { GetConsoleScreenBufferInfo(self.0.as_raw_handle(), info.as_mut_ptr()) };
-        if ret == 0 {
-            log::error!(
-                "GetConsoleScreenBufferInfo failed: {}",
-                io::Error::last_os_error()
-            );
-            return (0, 0);
-        }
-        let info = unsafe { info.assume_init() };
-
-        let cols = (info.srWindow.Right - info.srWindow.Left + 1) as u16;
-        let rows = (info.srWindow.Bottom - info.srWindow.Top + 1) as u16;
-        (cols, rows)
-    }
-}
-
-#[cfg(unix)]
-#[repr(C)]
-#[derive(Default)]
-struct WS {
-    rows: u16,
-    cols: u16,
-    xpixel: u16,
-    ypixel: u16,
-}
-
-#[cfg(unix)]
-ioctl_read_bad!(tiocgwinsz, TIOCGWINSZ, WS);
