+++ src/devices/src/virtio/console/port_io/mod.rs
@@ -1,0 +1,102 @@
+use log::Level;
+use std::io;
+use utils::eventfd::EventFd;
+use vm_memory::{VolatileSlice, WriteVolatile};
+
+#[cfg(unix)]
+mod unix;
+#[cfg(unix)]
+pub use unix::*;
+
+#[cfg(windows)]
+mod windows;
+#[cfg(windows)]
+pub use windows::*;
+
+pub trait PortInput {
+    fn read_volatile(&mut self, buf: &mut VolatileSlice) -> Result<usize, io::Error>;
+
+    fn wait_until_readable(&self, stopfd: Option<&EventFd>);
+}
+
+pub trait PortOutput {
+    fn write_volatile(&mut self, buf: &VolatileSlice) -> Result<usize, io::Error>;
+
+    fn wait_until_writable(&self);
+}
+
+/// Terminal properties associated with this port
+pub trait PortTerminalProperties: Send + Sync {
+    fn get_win_size(&self) -> (u16, u16);
+}
+
+pub fn term_fixed_size(width: u16, height: u16) -> Box<dyn PortTerminalProperties + Send + Sync> {
+    Box::new(PortTerminalPropertiesFixed((width, height)))
+}
+
+pub fn output_to_log_as_err() -> Box<dyn PortOutput + Send> {
+    Box::new(PortOutputLog::new())
+}
+
+struct PortTerminalPropertiesFixed((u16, u16));
+
+impl PortTerminalProperties for PortTerminalPropertiesFixed {
+    fn get_win_size(&self) -> (u16, u16) {
+        self.0
+    }
+}
+
+#[derive(Default)]
+pub struct PortOutputLog {
+    buf: Vec<u8>,
+}
+
+impl PortOutputLog {
+    const FORCE_FLUSH_TRESHOLD: usize = 512;
+    const LOG_TARGET: &'static str = "init_or_kernel";
+
+    fn new() -> Self {
+        Self::default()
+    }
+
+    fn force_flush(&mut self) {
+        log::log!(target: PortOutputLog::LOG_TARGET, Level::Error, "[missing newline]{}", String::from_utf8_lossy(&self.buf));
+        self.buf.clear();
+    }
+}
+
+impl PortOutput for PortOutputLog {
+    fn write_volatile(&mut self, buf: &VolatileSlice) -> Result<usize, io::Error> {
+        self.buf.write_volatile(buf).map_err(io::Error::other)?;
+
+        let mut start = 0;
+        for (i, ch) in self.buf.iter().cloned().enumerate() {
+            if ch == b'\n' {
+                log::log!(target: PortOutputLog::LOG_TARGET, Level::Error, "{}", String::from_utf8_lossy(&self.buf[start..i]));
+                start = i + 1;
+            }
+        }
+        self.buf.drain(0..start);
+        // Make sure to not grow the internal buffer forever!
+        if self.buf.len() > PortOutputLog::FORCE_FLUSH_TRESHOLD {
+            self.force_flush()
+        }
+        Ok(buf.len())
+    }
+
+    fn wait_until_writable(&self) {}
+}
+
+pub struct PortInputEmpty {}
+
+impl PortInputEmpty {
+    pub fn new() -> Self {
+        PortInputEmpty {}
+    }
+}
+
+impl Default for PortInputEmpty {
+    fn default() -> Self {
+        Self::new()
+    }
+}
